requirements.txt
google_api_python_client==2.120.0
google_auth_oauthlib==1.2.0
matplotlib==3.5.2
protobuf==4.25.3
selenium==4.18.1
